external: move external commands configuration to cmd package
Part of #1136
